Added support for re-athentication on Google
We needed to be able to re-authenticate on Google, and that meant that we
needed to add several infrastructural changes to the endpoints as well as
adding in a new 'defmethod' for the refreshing, and then we need to add in
the top-level function for refreshing and update the docs. This is all
tested out, and works like a charm.