Added the refresh-token BACK to token data
When we are doing a refresh of the OAuth2 token, we need to add it back
into the map to make sure that it's there and available for the next time
we might want to call this function. This was something I was doing in the
calling code, but it really belongs here as everyone is likely to need this.